nn/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
faiss-cpu>=1.7.4
opencv-python>=4.9.0
onnx>=1.16.0
onnxruntime>=1.18.0
numpy>=1.26.0
Pillow>=10.0.0
